+++ src/components/GridKeyboardNavigationContext/GridKeyboardNavigationContext.js
@@ -1,6 +1,6 @@
 import React, { useRef, useContext, useCallback } from "react";
 import useEventListener from "../../hooks/useEventListener";
-import { focusElementWithDirection, getDirectionMaps, getOppositeDirection, getOutmostElementInDirection } from "./GridKeyboardNavigationContextHelper";
+import { focusElementWithDirection, getDirectionMaps, getOppositeDirection, getOutmostElementInDirection } from "./helper";
 
 export const GridKeyboardNavigationContext = React.createContext();
